[SSHD-816] Provide some way for the SSHD command line server to detect and configure the (split) SFTP subsystem
