+++ sshd-cli/src/main/java/org/apache/sshd/cli/CliSupport.java
@@ -34,5 +34,4 @@ public abstract class CliSupport {
         stderr.println(message);
         return true;
     }
-
 }
